+++ src/test/java/picocli/RepeatableSubcommandsTest.java
@@ -701,6 +701,14 @@ public class RepeatableSubcommandsTest {
     }
 
     @Test
+    public void test1125_MultivalueOptions_ConsumesSubcommands() {
+        int exitCode = new CommandLine(new MultivalueTop())
+            .setAllowSubcommandsAsOptionParameters(true)
+            .execute("sub1 -x1 -x2 -x3 sub2 -y1 -y2 -y3 -y4".split(" "));
+        assertEquals(8, exitCode); // 8 args of the first sub1
+    }
+
+    @Test
     public void testCommandSpec_SubcommandsRepeatable() {
         class Positional {
             @Parameters(index = "0") String first;